+++ src/Schema/Grammars/GrammarIndex.php
@@ -40,4 +40,42 @@ trait GrammarIndex
     {
         return "alter table {$this->wrapTable($blueprint)} drop constraint if exists {$this->wrap($command->index)}";
     }
+
+    /**
+     * Compile a partial index key command.
+     */
+    public function compilePartialIndex(Blueprint $blueprint, Fluent $command): string
+    {
+        return sprintf('create index %s on %s%s (%s) where %s',
+            $this->wrap($command->index),
+            $this->wrapTable($blueprint),
+            $command->algorithm ? ' using '.$command->algorithm : '',
+            $this->columnize($command->columns),
+            $command->condition,
+        );
+    }
+
+    /**
+     * Compile a partial spatial index key command.
+     */
+    public function compilePartialSpatialIndex(Blueprint $blueprint, Fluent $command): string
+    {
+        $command->algorithm = 'gist';
+
+        return $this->compilePartialIndex($blueprint, $command);
+    }
+
+    /**
+     * Compile a partial unique key command.
+     */
+    public function compilePartialUnique(Blueprint $blueprint, Fluent $command): string
+    {
+        return sprintf('create unique index %s on %s%s (%s) where %s',
+            $this->wrap($command->index),
+            $this->wrapTable($blueprint),
+            $command->algorithm ? ' using '.$command->algorithm : '',
+            $this->columnize($command->columns),
+            $command->condition,
+        );
+    }
 }