+++ src/Support/Helpers/Query.php
@@ -2,24 +2,25 @@
 
 declare(strict_types=1);
 
-namespace Tpetry\PostgresqlEnhanced\Schema;
+namespace Tpetry\PostgresqlEnhanced\Support\Helpers;
 
+use Illuminate\Database\Connection;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use RuntimeException;
 
-trait BuilderSqlquery
+class Query
 {
     /**
      * Transforms a query to it's sql representation.
      */
-    public function makeSqlQuery(QueryBuilder | string $query): string
+    public static function toSql(QueryBuilder | string $query): string
     {
         if (\is_string($query)) {
             return $query;
         }
 
-        $bindings = $this->getConnection()->prepareBindings($query->getBindings());
-        $sql = preg_replace_callback('/(?<!\?)\?(?!\?)/', function () use (&$bindings) {
+        $bindings = $query->getConnection()->prepareBindings($query->getBindings());
+        $sql = preg_replace_callback('/(?<!\?)\?(?!\?)/', function () use (&$bindings, $query) {
             if (0 === \count($bindings)) {
                 throw new RuntimeException('Number of bindings does not match the number of placeholders');
             }
@@ -29,7 +30,7 @@ trait BuilderSqlquery
             return (string) match (true) {
                 null === $value => 'null',
                 is_numeric($value) => $value,
-                default => $this->getConnection()->getPdo()->quote((string) $value),
+                default => with($query->getConnection(), fn (Connection $connection) => $connection->getPdo()->quote((string) $value)),
             };
         }, $query->toSql());