+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - If exists index dropping
+- Partial indexes
 
 ## [0.2.1] - 2021-04-22
 ### Fixed